+++ src/cmds/init.rs
@@ -92,7 +92,8 @@ pub fn main(options: &cli::MainOptions, init_options: &mut InitOptions) -> Resul
     };
 
     let mut config = model::Configuration::new();
-    config.root = model::Variable::from_expr(init_options.root.clone());
+    config.root =
+        model::Variable::from_expr(constants::ROOT.to_string(), init_options.root.clone());
     config.root_path.clone_from(&dirname);
     config.path = Some(config_path.clone());